embed backend with frontend

+++ src/controllers/coops.js
@@ -8,18 +8,16 @@ class Coops {
 
 	// creating a function for add cooperative
 	static async createCoop(req,res){
-		req.headers[req.cookies];
-			console.log(req.headers);
+		// req.headers[req.cookies];
+		// 	console.log(req.headers);
 		// // try and catch to find if not exist and create new coop
 		try{ 
-			
+			console.log(req.user.userFind.id);
 			const {coopName,coopLocation,tin} = req.body;
-			const UserId= req.user.id;
+			const UserId= req.user.userFind.id;
 			const Op = Sequelize.Op;
 		// 	// find if exist
-			const findOne = await Coop.findOne({where: {
-				[Op.or]:[{tin},{UserId}]
-			}
+			const findOne = await Coop.findOne({where: {tin}
 			});
 			if (!findOne) {
 				const coudinary_links = await cloud(req.files);
@@ -40,14 +38,7 @@ class Coops {
 						result
 					})
 				})
-			}else{
-				return res.status(500).send({
-					status:res.statusCode,
-					message:'something went wrong!'
-					
-				})
 			}
-			
 		}
 		catch(err){
 			// res.status(500).send({
@@ -90,46 +81,61 @@ class Coops {
 	static async deleteCoop(req,res){
 		// check if tin is not empty
 		try{
-			const{tin}=req.body;
+			const {tin,coopName}= req.params;
+			const findOne = await Coop.findOne({where:{tin,coopName}});
+			if (findOne) {
 				await Coop.destroy({where:{tin}}).then(result=>{
 					if (result) {
-						return res.status(200).send({
-							status:res.statusCode,
+						return res.render('del-coop',{
 							message:'cooperative has been deleted successfully!',
-						})
-					}else{
-						res.status(500).send({
-							status:res.statusCode,
-							message:'cooperative with that tin not exist yet!'
 						})
 					}
 				})
+			} else{
+				return res.render('del-coop',{
+					message:`cooperative with tin ${tin} not exist !`,
+				})
+			}
+				
 			
 		}catch(err){
-			return res.status(500).send({
-				status:res.statusCode,
-				message:'something went wrong on server',
-			})
+			return res.redirect('/500');
 		}
 	}
 	// get all coops
 	static async getCoops(req,res){
 		try{	
 			await Coop.findAll().then(result =>{
-				return res.status(200).send({
-					status:res.statusCode,
-					message:'All cooperatives fetched successfully!',
+				return res.render('all-coops',{
 					result,
+					user:req.user.userFind
 				})
 			})
 		}catch(err){
-			res.status(500).send({
-				status:res.statusCode,
-				message:'something went wrong on server!'
-			})
+			res.redirect('/500');
 		}
 	}
-	// 
+	// get info of coop to delete
+	static async getDelInfo(req,res){
+		try{
+			const{coopName}=req.body;
+			const findOne = await Coop.findOne({where:{coopName}});
+			if (findOne) {
+				return res.render('del-coop',{
+					findOne,
+					user:req.user.userFind,
+					message:`${coopName}  found  successfully!`
+				})
+			} else {
+				return render('del-user',{
+					user:req.user.userFind,
+					message:`${coopName} not exist check cooperative name!`
+				})
+			}
+		}catch(err){
+			return res.redirect('/500');
+		}
+	}
 }
 
 // export coops
